# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  model.Song:
    properties:
      group:
        type: string
      id:
        type: string
      link:
        type: string
      release_date:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
  model.SongDTO:
    properties:
      group:
        type: string
      song:
        type: string
    type: object
info:
  contact: {}
  description: API for managing a song library
  title: Song Library API
  version: "1.0"
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Returns a list of all songs with optional filtering and pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Song ID
        in: query
        name: id
        type: string
      - description: Group name
        in: query
        name: group
        type: string
      - description: Song title
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Song'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to get library
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get all songs in the library
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Creates a new song with the given details
      parameters:
      - description: Song details
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/model.SongDTO'
      produces:
      - application/json
      responses:
        "200":
          description: song_id
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to create song
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create a new song
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Deletes a song with the given ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Song deleted successfully
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "400":
          description: Invalid song ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Record not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to delete song
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete a song
      tags:
      - songs
    put:
      consumes:
      - application/json
      description: Updates a song with the given ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated song details
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/model.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Song'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Record not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to update song
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Update an existing song
      tags:
      - songs
  /songs/{id}/verses:
    get:
      consumes:
      - application/json
      description: Returns paginated verses for the specified song
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of verses per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Invalid song ID or pagination parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to get song verses
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get song verses
      tags:
      - songs
swagger: "2.0"